feat!: BREAKING Make WebVitals a plugin instead of an optional property on Portal
